convert example_height into interger

+++ python-modified/ex3/displayData.py
@@ -28,7 +28,7 @@ def displayData(X, example_width=None):
 
 	# Compute rows, cols
 	m, n = X.shape
-	example_height = n / example_width
+	example_height = int(n / example_width)
 
 	# Compute number of items to display
 	display_rows = int(math.floor(math.sqrt(m)))
@@ -67,7 +67,8 @@ def displayData(X, example_width=None):
 			#	then the second column, etc. This latter behaviour is what we want.
 			#	Alternatively, we can keep the deault order="C" and then transpose the result
 			#	from the reshape call.
-			display_array[rows[0]:rows[-1]+1 , cols[0]:cols[-1]+1] = np.reshape(X[curr_ex-1, :], (example_height, example_width), order="F") / max_val
+			display_array[rows[0]:rows[-1]+1 , cols[0]:cols[-1]+1] = np.reshape(X[curr_ex-1, :],
+								(example_height, example_width), order="F") / max_val
 			curr_ex += 1
 
 		if curr_ex > m:
